fix:mybatis-plus版本升级处理

+++ src/main/java/com/liangzhicheng/modules/service/impl/SysUserServiceImpl.java
@@ -312,9 +312,9 @@ public class SysUserServiceImpl extends ServiceImpl<ISysUserDao, SysUserEntity> 
                 throw new CustomizeException(ApiConstant.BASE_FAIL_CODE, "账号名称字数过长");
             }
             if(SysToolUtil.isBlank(id) || !user.getAccountName().equals(accountName)){
-                Integer count = baseMapper.selectCount(new QueryWrapper<SysUserEntity>()
+                Long count = baseMapper.selectCount(new QueryWrapper<SysUserEntity>()
                         .eq("account_name", accountName).eq(Constants.DEL_FLAG, Constants.ZERO));
-                if(count > 0){
+                if(count.intValue() > 0){
                     throw new CustomizeException(ApiConstant.BASE_FAIL_CODE, "账号已存在");
                 }
             }
